.history/requirements_20250810150313.txt
pip install openai>=1.30.0 \
            langchain \
            langchain-community \
            langchain-openai \
            faiss-cpu \
            openai-whisper \
            tiktoken \
            matplotlib \
            python-dotenv \
            pyttsx3
